+++ examples/todos/lib/todos_page.dart
@@ -1,30 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_solidart/flutter_solidart.dart';
 import 'package:todos/controllers/controller.dart';
 import 'package:todos/models/todo.dart';
 import 'package:todos/widgets/todos_body.dart';
 
-class TodosPage extends StatefulWidget {
+class TodosPage extends StatelessWidget {
   const TodosPage({super.key});
 
   @override
-  State<TodosPage> createState() => _TodosPageState();
-}
-
-class _TodosPageState extends State<TodosPage> {
-  final todosController = TodosController(initialTodos: Todo.sample);
-
-  @override
-  void dispose() {
-    todosController.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    // Using Provider here to provide the [TodosController] to descendants.
-    return Provider(
-      create: (_) => todosController,
+    // Using SolidProvider here to provide the [TodosController] to descendants.
+    return Solid(
+      providers: [
+        SolidProvider<TodosController>(
+          create: () => TodosController(initialTodos: Todo.sample),
+          dispose: (controller) => controller.dispose(),
+        ),
+      ],
       child: Scaffold(
         appBar: AppBar(title: const Text('Todos')),
         body: const Padding(